Make a good demo page to show the usage better
Fix a issue because focus was causing a issue when using the increase/decrease buttons
Fix a Nan issue when the input is a number and the value is empty (NaN)
Use MIN/MAX_SAFE_INTEGER instead.
Fix an issue where if we are 0 or 0.000 (float/decimal), that we would have to clear the input box before we could start clicking numbers.